Add support for markdown tables, along with a custom renderer that adds class "table" so bootstrap will style them.
